# api/openapi/accounts.yml
openapi: 3.0.3
info:
  title: Pinjamean Identity API
  version: "1.0"

paths:
  /accounts:
    post:
      summary: Save an account
      operationId: saveAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAccountRequest'
      responses:
        '204':
          description: Account saved successfully
        '400':
          description: Invalid account data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{email}:
    get:
      summary: Get account by email
      operationId: getAccountByEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    AccountRole:
      type: string
      enum:
        - officer
        - credit_analyst

    SaveAccountRequest:
      type: object
      required:
        - uuid
        - email
        - password
        - role
      properties:
        uuid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
        role:
          $ref: '#/components/schemas/AccountRole'

    AccountResponse:
      type: object
      required:
        - uuid
        - email
        - role
      properties:
        uuid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/AccountRole'

    ErrorResponse:
      type: object
      required:
        - slug
        - message
      properties:
        slug:
          type: string
        message:
          type: string