# internal/identity/ports/account.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package ports

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for AccountRole.
const (
	CreditAnalyst AccountRole = "credit_analyst"
	Officer       AccountRole = "officer"
)

// Valid indicates whether the value is a known member of the AccountRole enum.
func (e AccountRole) Valid() bool {
	switch e {
	case CreditAnalyst:
		return true
	case Officer:
		return true
	default:
		return false
	}
}

// AccountResponse defines model for AccountResponse.
type AccountResponse struct {
	Email openapi_types.Email `json:"email"`
	Role  AccountRole         `json:"role"`
	Uuid  openapi_types.UUID  `json:"uuid"`
}

// AccountRole defines model for AccountRole.
type AccountRole string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// SaveAccountRequest defines model for SaveAccountRequest.
type SaveAccountRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password *string             `json:"password,omitempty"`
	Role     AccountRole         `json:"role"`
	Uuid     openapi_types.UUID  `json:"uuid"`
}

// SaveAccountJSONRequestBody defines body for SaveAccount for application/json ContentType.
type SaveAccountJSONRequestBody = SaveAccountRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// SaveAccount Save an account
	// (POST /accounts)
	SaveAccount(w http.ResponseWriter, r *http.Request)
	// GetAccountByEmail Get account by email
	// (GET /accounts/{email})
	GetAccountByEmail(w http.ResponseWriter, r *http.Request, email openapi_types.Email)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// SaveAccount Save an account
// (POST /accounts)
func (_ Unimplemented) SaveAccount(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetAccountByEmail Get account by email
// (GET /accounts/{email})
func (_ Unimplemented) GetAccountByEmail(w http.ResponseWriter, r *http.Request, email openapi_types.Email) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// SaveAccount operation middleware
func (siw *ServerInterfaceWrapper) SaveAccount(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SaveAccount(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetAccountByEmail operation middleware
func (siw *ServerInterfaceWrapper) GetAccountByEmail(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "email" -------------
	var email openapi_types.Email

	err = runtime.BindStyledParameterWithOptions("simple", "email", chi.URLParam(r, "email"), &email, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "email", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "email", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAccountByEmail(w, r, email)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/accounts", wrapper.SaveAccount)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/accounts/{email}", wrapper.GetAccountByEmail)
	})

	return r
}

type SaveAccountRequestObject struct {
	Body *SaveAccountJSONRequestBody
}

type SaveAccountResponseObject interface {
	VisitSaveAccountResponse(w http.ResponseWriter) error
}

type SaveAccount204Response struct {
}

func (response SaveAccount204Response) VisitSaveAccountResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type SaveAccount400JSONResponse ErrorResponse

func (response SaveAccount400JSONResponse) VisitSaveAccountResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetAccountByEmailRequestObject struct {
	Email openapi_types.Email `json:"email"`
}

type GetAccountByEmailResponseObject interface {
	VisitGetAccountByEmailResponse(w http.ResponseWriter) error
}

type GetAccountByEmail200JSONResponse AccountResponse

func (response GetAccountByEmail200JSONResponse) VisitGetAccountByEmailResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetAccountByEmail404JSONResponse ErrorResponse

func (response GetAccountByEmail404JSONResponse) VisitGetAccountByEmailResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// SaveAccount Save an account
	// (POST /accounts)
	SaveAccount(ctx context.Context, request SaveAccountRequestObject) (SaveAccountResponseObject, error)
	// GetAccountByEmail Get account by email
	// (GET /accounts/{email})
	GetAccountByEmail(ctx context.Context, request GetAccountByEmailRequestObject) (GetAccountByEmailResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// SaveAccount operation middleware
func (sh *strictHandler) SaveAccount(w http.ResponseWriter, r *http.Request) {
	var request SaveAccountRequestObject

	var body SaveAccountJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.SaveAccount(ctx, request.(SaveAccountRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "SaveAccount")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(SaveAccountResponseObject); ok {
		if err := validResponse.VisitSaveAccountResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetAccountByEmail operation middleware
func (sh *strictHandler) GetAccountByEmail(w http.ResponseWriter, r *http.Request, email openapi_types.Email) {
	var request GetAccountByEmailRequestObject

	request.Email = email

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetAccountByEmail(ctx, request.(GetAccountByEmailRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetAccountByEmail")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetAccountByEmailResponseObject); ok {
		if err := validResponse.VisitGetAccountByEmailResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}